+++ app.vue
@@ -27,7 +27,6 @@ useHead(() => ({
   },
   link: [
     ...(head.value.link || []),
-    { rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' },
     { rel: 'icon', type: 'image/svg+xml', href: '/favicon.svg' }
   ],
   meta: [