+++ components/UiIcon.vue
@@ -1,0 +1,77 @@
+<template>
+  <svg
+    :width="size"
+    :height="size"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    stroke-width="2"
+    stroke-linecap="round"
+    stroke-linejoin="round"
+    aria-hidden="true"
+    class="ui-icon"
+  >
+    <path v-for="(d, i) in paths" :key="i" :d="d" />
+  </svg>
+</template>
+
+<script setup>
+import { computed } from 'vue'
+
+// Lucide-style inline icon set (ISC license, lucide.dev)
+const ICONS = {
+  zap: ['M13 2 3 14h9l-1 8 10-12h-9l1-8z'],
+  shield: ['M20 13c0 5-3.5 7.5-7.66 8.95a1 1 0 0 1-.67-.01C7.5 20.5 4 18 4 13V6a1 1 0 0 1 1-1c2 0 4.5-1.2 6.24-2.72a1.17 1.17 0 0 1 1.52 0C14.51 3.81 17 5 19 5a1 1 0 0 1 1 1z'],
+  message: ['M7.9 20A9 9 0 1 0 4 16.1L2 22Z'],
+  rocket: [
+    'M4.5 16.5c-1.5 1.26-2 5-2 5s3.74-.5 5-2c.71-.84.7-2.13-.09-2.91a2.18 2.18 0 0 0-2.91-.09z',
+    'm12 15-3-3a22 22 0 0 1 2-3.95A12.88 12.88 0 0 1 22 2c0 2.72-.78 7.5-6 11a22.35 22.35 0 0 1-4 2z',
+    'M9 12H4s.55-3.03 2-4c1.62-1.08 5 0 5 0',
+    'M12 15v5s3.03-.55 4-2c1.08-1.62 0-5 0-5'
+  ],
+  code: ['m16 18 6-6-6-6', 'm8 6-6 6 6 6'],
+  atom: [
+    'M12 12h.01',
+    'M20.2 20.2c2.04-2.03.02-7.36-4.5-11.9-4.54-4.52-9.87-6.54-11.9-4.5-2.04 2.03-.02 7.36 4.5 11.9 4.54 4.52 9.87 6.54 11.9 4.5Z',
+    'M15.7 15.7c4.52-4.54 6.54-9.87 4.5-11.9-2.03-2.04-7.36-.02-11.9 4.5-4.52 4.55-6.54 9.87-4.5 11.9 2.03 2.04 7.36.02 11.9-4.5Z'
+  ],
+  users: [
+    'M16 21v-2a4 4 0 0 0-4-4H6a4 4 0 0 0-4 4v2',
+    'M13 7a4 4 0 1 1-8 0 4 4 0 0 1 8 0z',
+    'M22 21v-2a4 4 0 0 0-3-3.87',
+    'M16 3.13a4 4 0 0 1 0 7.75'
+  ],
+  mail: [
+    'M4 4h16c1.1 0 2 .9 2 2v12c0 1.1-.9 2-2 2H4c-1.1 0-2-.9-2-2V6c0-1.1.9-2 2-2z',
+    'm22 6-10 7L2 6'
+  ],
+  'book-open': [
+    'M12 7v14',
+    'M3 18a1 1 0 0 1-1-1V4a1 1 0 0 1 1-1h5a4 4 0 0 1 4 4 4 4 0 0 1 4-4h5a1 1 0 0 1 1 1v13a1 1 0 0 1-1 1h-6a3 3 0 0 0-3 3 3 3 0 0 0-3-3z'
+  ],
+  layers: [
+    'M12 2 2 7l10 5 10-5-10-5z',
+    'm2 17 10 5 10-5',
+    'm2 12 10 5 10-5'
+  ],
+  'arrow-right': ['M5 12h14', 'm12 5 7 7-7 7'],
+  sparkles: [
+    'M9.937 15.5A2 2 0 0 0 8.5 14.063l-6.135-1.582a.5.5 0 0 1 0-.962L8.5 9.936A2 2 0 0 0 9.937 8.5l1.582-6.135a.5.5 0 0 1 .963 0L14.063 8.5A2 2 0 0 0 15.5 9.937l6.135 1.581a.5.5 0 0 1 0 .964L15.5 14.063a2 2 0 0 0-1.437 1.437l-1.582 6.135a.5.5 0 0 1-.963 0z'
+  ]
+}
+
+const props = defineProps({
+  icon: { type: String, required: true },
+  size: { type: Number, default: 24 }
+})
+
+const paths = computed(() => ICONS[props.icon] || [])
+</script>
+
+<style scoped>
+.ui-icon {
+  display: inline-block;
+  vertical-align: middle;
+  flex: none;
+}
+</style>